chore(lint): consolidate test-fixture string literals into constants
CI's goconst surfaced 7 repeated literals in bootstrap_resources_test.go
("default" 48×, "pacific-1" 19×, etc.). Add testNamespaceDefault,
testChainPacific, testChainAtlantic, testImageLatest, testValidatorName,
testExporterName, testDataVolumeName and replace the literal occurrences.

Also drops two unconvert hits — seiconfig.PortSidecar is already int32.

+++ internal/task/bootstrap_node_adapter.go
@@ -32,7 +32,7 @@ func nodeToBootstrapInputs(node *seiv1alpha1.SeiNode, snap *seiv1alpha1.Snapshot
 		sidecarImage = node.Spec.Sidecar.Image
 	}
 
-	sidecarPort := int32(seiconfig.PortSidecar)
+	sidecarPort := seiconfig.PortSidecar
 	if node.Spec.Sidecar != nil && node.Spec.Sidecar.Port != 0 {
 		sidecarPort = node.Spec.Sidecar.Port
 	}
